# openapi.yaml
openapi: 3.1.0
info:
  title: League Table API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: |
    The backend the League Table frontend (`frontend/services/httpApi.js`) talks to.

    Scoring, the commitments adjustment, ranking and tie-breaking all run on the server;
    the browser only renders the result. A student's raw score is weighted by criterion,
    multiplied by a capped factor from the weekly hours they report and approved by the
    instructor, and clipped at 100. Classmates see only the final adjusted score and rank.

    **Conventions**
    - Every field, query parameter and path parameter is `snake_case`, matching the
      data model in `_docs/spec.md` and FastAPI's defaults. The frontend's services
      layer translates its own JS option names (`weekId`, `windowMode`, ...) at the
      HTTP boundary.
    - Timestamps (`last_updated`, `submitted_at`, ...) are ISO 8601 UTC strings, or
      `null` before any successful read.
    - Errors carry a human-readable `detail` message and an HTTP status; the client
      shows the message to the user.

    **Authentication**
    Every endpoint except `POST /auth/login` and `GET /demo/accounts` needs a session
    cookie. Roles are enforced by the server, never the browser. An `instructor` can do
    everything; a `student` may read the full ranking but only their own breakdown and
    commitments, and may not change settings or approve anything. Sign-in is by access
...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /classes/{class_id}/ranking:
    get:
      tags: [Ranking]
      operationId: getRanking
      summary: The ranked table for a week or cumulative window
      description: |
        Scores every student with entries in the window and ranks them by adjusted score.
        Each week is scored on its own: criteria are weighted into a raw score, multiplied
        by that week's commitments factor and clipped at 100. A window of several weeks
        averages the weekly adjusted scores, and a week where a student has no entries is
        left out of their average. Students level on adjusted score are ordered by raw
        score, then by the tie-breaker criteria (attendance, then homework); any still level
        share a rank (1, 1, 3). A criterion a student has no data for is left out and the
        remaining weights are rescaled, so missing data is never scored as zero.

        Rows carry the adjusted score and rank only: never another student's raw score,
        factor or hours.

        `rank_delta` compares against the previous week's table with the same
        criteria, weights and window mode, and is `null` for the first week or when
        the student was not ranked then.

        Any signed-in role may call this. `weights` lets the instructor preview weights
        before saving them; when omitted, the saved weights are used.
      parameters:
        - $ref: '#/components/parameters/ClassId'
        - $ref: '#/components/parameters/Week'
        - $ref: '#/components/parameters/Window'
        - $ref: '#/components/parameters/RollingWeeks'
        - $ref: '#/components/parameters/Criteria'
        - $ref: '#/components/parameters/Weights'
        - $ref: '#/components/parameters/NameMode'
      responses:
        '200':
          description: The ranking, best first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ranking'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/students/{student_id}/explanation:
    get:
      tags: [Ranking]
      operationId: getExplanation
      summary: How one student's rank was reached, criterion by criterion
      description: |
        The instructor may open any student. A student may open only their own; asking
        for another student's breakdown returns 403 regardless of what the client
        sends. Accepts the same window, criteria, weights and name mode as the
        ranking so the breakdown always matches the table being viewed.

        Shows the criteria (which add up to the raw score), the weighted hours by type,
        the factor, the adjusted score, whether it was clipped at 100 and the gap to the
        next rank. For a window of several weeks, `weeks` lists each week and the scores
        are means.
      parameters:
        - $ref: '#/components/parameters/ClassId'
        - name: student_id
          in: path
          required: true
          schema:
            type: string
          example: s01
        - $ref: '#/components/parameters/Week'
        - $ref: '#/components/parameters/Window'
        - $ref: '#/components/parameters/RollingWeeks'
        - $ref: '#/components/parameters/Criteria'
        - $ref: '#/components/parameters/Weights'
        - $ref: '#/components/parameters/NameMode'
      responses:
        '200':
          description: The breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Explanation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Unknown class or student, or the student has no entries in this window.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
  /classes/{class_id}/status:
    get:
      tags: [Data]
      operationId: getStatus
      summary: Data freshness and validation issues
      description: |
        Read from the server's cache; never triggers a fetch from the data source.
        `last_updated` is `null` before the first successful read.
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Current status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /classes/{class_id}/refresh:
    post:
      tags: [Data]
      operationId: refresh
      summary: Re-read the data source now (instructor only)
      description: |
        Bypasses the server's cache (60 s TTL). Rate limited to one call per 10
        seconds; a call inside the cooldown returns 429 with a message such as
        "Please wait 7s before refreshing again." If the source fails but a previous
        snapshot exists, the call succeeds with `stale: true`.
      parameters:
        - $ref: '#/components/parameters/ClassId'
      responses:
        '200':
          description: Refreshed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Refreshed too recently.
          headers:
            Retry-After:
              description: Seconds until the next refresh is allowed.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/SourceUnavailable'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `/auth/login`.
  parameters:
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
      example: c1
    Week:
      name: week
      in: query
      required: true
      description: ID of the week to rank. With `window=cumulative`, this is the last week included.
      schema:
        type: string
      example: w5
    Window:
      name: window
      in: query
      description: |
        `week` scores only that week. `cumulative` averages the weekly adjusted scores from
        week 1 through `week`, and `rolling` averages the last `rolling_weeks` weeks ending
        at `week`.
      schema:
        type: string
        enum: [week, cumulative, rolling]
        default: week
    RollingWeeks:
      name: rolling_weeks
      in: query
      description: How many weeks a `rolling` window covers.
      schema:
        type: integer
        minimum: 2
        maximum: 16
        default: 4
...
    Criteria:
      name: criteria
      in: query
      description: Criterion keys to include, comma-separated. Omitted or empty means all criteria.
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
      example: [homework, attendance]
    Weights:
      name: weights
      in: query
      description: |
        JSON-encoded object mapping criterion key to weight, for previewing unsaved
        weights. Criteria not listed get weight 0. The server rescales the weights
        of the selected criteria to total 100. Omitted means use the saved weights.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Weights'
          example: { homework: 40, attendance: 30, participation: 30 }
    NameMode:
      name: name_mode
      in: query
      description: |
        How names are shown in `display_name`. `nickname` falls back to the full
        name when a student has none; `initials` shows "First L.".
      schema:
        type: string
        enum: [full, nickname, initials]
        default: full
  responses:
    Unauthorized:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The caller's role does not allow this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: A student may only open their own breakdown.
    NotFound:
      description: Unknown class.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: |
        A parameter or body field is missing or malformed. Uses FastAPI's default 422
        body, where `detail` is a list of per-field problems.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    SourceUnavailable:
      description: |
        The data source did not respond and there is no earlier snapshot to serve.
        When a snapshot exists the server serves it with `stale: true` instead.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: The data source did not respond.
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Message safe to show to the user.
    ValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            type: object
            required: [loc, msg, type]
            properties:
              loc:
                type: array
                description: Where the problem is, e.g. `["query", "week"]`.
                items:
                  type: [string, integer]
              msg:
                type: string
              type:
                type: string
...
    Weights:
      type: object
      description: Criterion key to weight. Values are non-negative.
      additionalProperties:
        type: number
        minimum: 0
      example: { homework: 25, attendance: 20, participation: 20, project: 25, quizzes: 10 }
...
    SourceInfo:
      type: object
      required: [kind, label, demo]
      properties:
        kind:
          type: string
          description: Which data-source adapter is active (`toy` for seeded demo data).
          example: toy
        label:
          type: string
          example: Demo data
        demo:
          type: boolean
          description: True when the data is seeded demo data, so the UI can say so.
    Issue:
      type: object
      description: A problem found while validating the source data. The table is still served.
      required: [level, where, message]
      properties:
        level:
          type: string
          enum: [error, warning]
        where:
          type: string
          description: ID of the offending entry.
          example: e12
        message:
          type: string
          example: Unknown student ID "s99"
...
    RankingRow:
      type: object
      required: [student_id, display_name, score, rank, tied, rank_delta, missing, gap_to_next, gap_to_below]
      properties:
        student_id:
          type: string
        display_name:
          type: string
          description: Name in the requested `name_mode`.
        score:
          type: number
          description: The adjusted score, 0 to 100. Nothing else about a classmate's score is sent.
        rank:
          type: integer
          minimum: 1
          description: Competition rank; students level on every tie-breaker share it.
        tied:
          type: boolean
          description: True for every student in a group that shares a rank, including the first.
        rank_delta:
          type: [integer, 'null']
          description: Places moved since the previous week (positive is up); `null` if there is no previous rank.
        missing:
          type: array
          description: Keys of weighted criteria the student has no data for in this window.
          items:
            type: string
        gap_to_next:
          type: [number, 'null']
          description: Points needed to reach the next rank up; `null` at the top.
        gap_to_below:
          type: [number, 'null']
          description: Lead over the next rank down; `null` at the bottom.
    Ranking:
      type: object
      required: [class_id, week_id, window_mode, rolling_weeks, criteria_keys, weights, rows, last_updated, stale, issues]
      properties:
        class_id:
          type: string
        week_id:
          type: string
        window_mode:
          type: string
          enum: [week, cumulative, rolling]
        rolling_weeks:
          type: [integer, 'null']
          description: The window length for `rolling`; `null` otherwise.
        criteria_keys:
          type: array
          description: The criteria actually used (all criteria when none were requested).
          items:
            type: string
        weights:
          $ref: '#/components/schemas/Weights'
        rows:
          type: array
          description: Best first. Only students with entries in the window appear.
          items:
            $ref: '#/components/schemas/RankingRow'
        last_updated:
          type: string
          format: date-time
          description: When the source was last read successfully (ISO 8601, UTC).
        stale:
          type: boolean
          description: True when the source failed and the last good snapshot was served instead.
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
    ScorePart:
      type: object
      description: One criterion's contribution to a student's score.
      required: [key, label, earned, possible, normalised, weight, effective_weight, points, missing]
      properties:
        key:
          type: string
        label:
          type: string
        earned:
          type: [number, 'null']
          description: Sum of points earned in the window; `null` when missing.
        possible:
          type: [number, 'null']
          description: Sum of points available in the window; `null` when missing.
        normalised:
          type: [number, 'null']
          description: '`earned / possible` as 0 to 100; `null` when missing.'
        weight:
          type: number
          description: Weight as requested.
        effective_weight:
          type: number
          description: Weight after dropping missing criteria and rescaling to 100; 0 when missing.
        points:
          type: number
          description: '`normalised × effective_weight / 100`; 0 when missing. Parts sum to `raw_score`.'
        missing:
          type: boolean
          description: No data for this criterion in the window; it is excluded, not scored as zero.
    WeekBreakdown:
      type: object
      description: One week of a student's score, before and after the adjustment.
      required: [week_id, week_number, raw_score, weighted_hours, factor, adjusted_score, capped]
      properties:
        week_id:
          type: string
        week_number:
          type: integer
        raw_score:
          type: number
        weighted_hours:
          type: number
        factor:
          type: number
        adjusted_score:
          type: number
        capped:
          type: boolean
          description: The adjusted score was clipped at 100.
    Explanation:
      type: object
      description: |
        `score` is the adjusted score the table ranks on. For a window of several weeks each week
        is scored on its own and the weekly scores are averaged, so `raw_score` and `score` are
        means, `parts` average the weekly points, and `factor`, `weighted_hours` and `hours` are
        `null`; see `weeks`.
      required: [student_id, display_name, rank, tied, score, raw_score, factor, weighted_hours, hours, capped, parts, weeks, gap_to_next, gap_to_below, above, week_id, window_mode, rolling_weeks]
      properties:
        student_id:
          type: string
        display_name:
          type: string
        rank:
          type: integer
          minimum: 1
        tied:
          type: boolean
        score:
          type: number
          description: The adjusted score.
        raw_score:
          type: number
          description: The weighted criteria before the factor; the parts add up to it.
        factor:
          type: [number, 'null']
        weighted_hours:
          type: [number, 'null']
        hours:
          oneOf:
            - $ref: '#/components/schemas/Hours'
            - type: 'null'
          description: Weighted hours by commitment type, for a single week.
        capped:
          type: boolean
          description: Any week's adjusted score was clipped at 100.
        parts:
          type: array
          description: One per selected criterion, in criterion order.
          items:
            $ref: '#/components/schemas/ScorePart'
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/WeekBreakdown'
        gap_to_next:
          type: [number, 'null']
        gap_to_below:
          type: [number, 'null']
        above:
          type: [string, 'null']
          description: Display name of the nearest student ranked above; `null` at the top.
        week_id:
          type: string
        window_mode:
          type: string
          enum: [week, cumulative, rolling]
        rolling_weeks:
          type: [integer, 'null']
    Status:
      type: object
      required: [last_updated, stale, issues, source]
      properties:
        last_updated:
          type: [string, 'null']
          format: date-time
          description: ISO 8601 UTC time of the last successful read, or `null` before the first.
        stale:
          type: boolean
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
        source:
          $ref: '#/components/schemas/SourceInfo'
    RefreshResult:
      type: object
      required: [last_updated, issues, stale]
      properties:
        last_updated:
          type: string
          format: date-time
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
        stale:
          type: boolean
    Hours:
      type: object
      description: Hours per week by commitment type key, such as `work`, `childcare` and `eldercare`.
      additionalProperties:
        type: number
      example: { work: 12, childcare: 6, eldercare: 0 }
...